requirements.txt
asgiref==3.8.1
bitrix24-rest==1.1.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2
django-debug-toolbar==4.0.0
django-environ==0.11.2
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
idna==3.10
pillow==10.2.0
psycopg2==2.9.9
python-magic==0.4.27
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
django-cors-headers==4.9.0
sorl-thumbnail==13.0.0
djangorestframework==3.16.1
django-filter==23.2
Markdown==3.10.2